Fixed folder structure
PCOR-8

Fixed initial folder structure to make use of cmake features instead of relying on scripts
