--- dataprovider/CMakeLists.txt
@@ -1,2 +1,0 @@
-cmake_minimum_required(VERSION 3.9.2)
-project(dataprovider)
