+++ domain/public/views/authviewif.hpp
@@ -9,8 +9,7 @@
 **************************************************************************************************/
 #ifndef DOMAIN_PUBLIC_AUTHVIEWIF_HPP_
 #define DOMAIN_PUBLIC_AUTHVIEWIF_HPP_
-
-#include <inc/employee.hpp>
+#include <string>
 
 namespace domain {
 namespace authentication {
@@ -20,9 +19,21 @@ class AuthViewIface {
     AuthViewIface() = default;
     virtual ~AuthViewIface() = default;
 
+    /**
+    * showLoginScreen
+    */
     virtual void showLoginScreen() = 0;
+
+    /**
+    * showLoginFailed
+    */
     virtual void showLoginFailed() = 0;
-    virtual void loginSuccessful(const entities::Employee& employee) = 0;
+    
+    /**
+    * Will be called if user was found
+    * @param employee a JSON string containing employee details
+    */
+    virtual void loginSuccessful(const std::string& employee) = 0;
 };
 
 }  // namespace authentication
