+++ domain/src/controllers/authcontroller.cpp
@@ -7,19 +7,29 @@
 *                   Written by Ben Ziv <pointonsoftware@gmail.com>, August 2020                   *
 *                                                                                                 *
 **************************************************************************************************/
-#include <inc/employee.hpp>
+#include <controllers/authcontroller.hpp>
 
 namespace domain {
-namespace entities {
+namespace authentication {
 
-Employee::Employee(const std::string& firstname,
-                   const std::string& middlename,
-                   const std::string& lastname,
-                   const std::string& birthdate,
-                   const std::string& gender)
-: Person{firstname, middlename, lastname, birthdate, gender} {
+AuthController::AuthController(std::unique_ptr<AuthViewIface>&& view,
+                               std::unique_ptr<AuthDataProviderIface>&& dataprovider)
+: mView(std::move(view)), mDataProvider(std::move(dataprovider)) {
     // Empty for now
 }
 
-}  // namespace entities
+bool AuthController::login(const std::string& pin) {
+    // Check if pin is empty
+    // Check if pin is numeric
+    // Call authenticate
+    return true;
+}
+
+status::General AuthController::authenticate(const std::string& pin) {
+    // Check if dataprovider is ready; else throw
+    // Check pin in dataprovider
+    return status::General::SUCCESS;
+}
+
+}  // namespace authentication
 }  // namespace domain
