+++ domain/unittest/CMakeLists.txt
@@ -1,0 +1,1 @@
+project (domain_unittest)
