--- scripts/clean.sh
@@ -1,11 +1,0 @@
-#!/bin/sh
-
-# Clean bins
-cd $BIN_DIR
-ls -Q | grep -v .gitignore | xargs rm -rf
-
-# Clean libs
-cd $LIB_DIR
-ls -Q | grep -v .gitignore | xargs rm -rf
-
-echo "Cleaned!"
